CDS_Warm_up
The CDS Warm-Up repository is designed as an entry point for users to familiarize themselves with seasonal forecast data provided by the Copernicus Climate Change Service (C3S). As a component of a broader initiative focusing on the use of Copernicus data for impact assessment through CLIMADA, it presents an initial collection of exercises and tutorials that guide users in accessing, processing, and analyzing Copernicus data. The aim is to introduce users to data and impact assessment models, providing them with the knowledge to carry out impact analysis of seasonal forecasting data.